Issue #60 Add debug function
